requirements.txt
pandas>=2.0
plotly>=5.0
streamlit>=1.30
Pillow>=10.0
matplotlib>=3.8
